move pageNav component to shared location
